genai-risk-advisor/backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyMuPDF
faiss-cpu
langchain
langchain-groq
sentence-transformers
